requirement.txt
cffi==1.11.2
Django==1.11.5
django-bootstrap3==9.1.0
django-braces==1.11.0
djangorestframework==3.6.2
Faker==0.8.6
humanize==0.5.1
misaka==2.1.0
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
text-unidecode==1.0
